requirements.txt
aiofiles==23.2.1
aiohttp==3.9.4
chat_exporter==2.7.1
discord.py==2.3.2
disnake==2.9.1
fuzzywuzzy==0.18.0
numba==0.59.1
numpy==1.26.4
pydub==0.25.1
Requests==2.31.0
scikit_learn==1.4.2
scipy==1.13.0
SpeechRecognition==3.10.3
youtube_dl==2021.12.17
python-Levenshtein==0.25.1 
pillow==10.4.0
python-dotenv==1.0.1
aiosqlite==0.20.0